requirements.txt
python-dotenv
mongoengine~=0.27.0
pymongo~=4.3.3
pydantic-settings
PyYAML
uvicorn
python-multipart
fastapi~=0.115.6
openai
pydantic
redis
websockets
rq
bcrypt
requests
fireworks-ai
botocore
boto3
